Simplify the repository structure
